requirements.txt
robotframework==7.1
